pyproject.toml
[tool.poetry]
name = "radiology-gpt"
version = "0.1.0"
description = ""
authors = ["putu-adi-hyrd <putu.adi@hyrd.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<3.9.7 || >3.9.7,<3.12"
pandas = "^2.1.4"
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
torch = "^2.1.2"
torchvision = "^0.16.2"
timm = "^0.9.12"
h5py = "^3.10.0"
langchain = "^0.1.0"
transformers = "^4.36.2"
chromadb = "^0.4.22"
unstructured = "^0.12.0"
pdf2image = "^1.17.0"
openai = "^1.7.2"
streamlit = "^1.30.0"
langchain-openai = "^0.0.2.post1"
boto3 = "^1.34.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
